pipeline/requirements.txt

pandas==2.2.2
numpy==1.26.4
gspread==6.1.2
google-auth==2.33.0
